src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/getComputedStyle-flat-tree-parent-dirty-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: document.getElementById(...).createShadowRoot is not a function
getComputedStyle should update style if the parent node in the flat tree needs a recalc.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
